requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
MySQL-python==1.2.5
pytz==2017.2
setupfiles==0.0.50
